requirements.txt
matplotlib==3.5.3
numpy==1.21.6
Pillow==9.5.0
Pillow==11.0.0
scikit_learn==1.0.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.67.0
